+++ ocdskingfishercolab/__init__.py
@@ -5,20 +5,37 @@ from urllib.parse import urljoin
 
 import flattentool
 import gspread
-import pandas
-import psycopg2
 import requests
+import sql
 from google.colab import auth, files
 from gspread_dataframe import set_with_dataframe
+from IPython import get_ipython
 from libcoveocds.config import LibCoveOCDSConfig
 from notebook import notebookapp
 from oauth2client.client import GoogleCredentials
-from psycopg2.sql import SQL, Identifier
 from pydrive.auth import GoogleAuth
 from pydrive.drive import GoogleDrive
 
+
+# Monkeypatch ipython-sql's sql run function, to add a comment linking to the
+# colab notebook that it's run from
+# We monkeypatch run(), but don't call it directly below, as calling the magic
+# will handle connections for us.
+old_run = sql.run.run
+
+
+def run(conn, _sql, *args, **kwargs):
+    try:
+        comment = '/* https://colab.research.google.com/drive/{} */'.format(_notebook_id())
+    except KeyError:
+        comment = "/* run from a notebook, but no colab id */"
+    return old_run(conn, comment + _sql, *args, **kwargs)
+
+
+sql.run.run = run
+
+
 spreadsheet_name = None
-conn = None
 
 # Use the same placeholder values as OCDS Kit.
 package_metadata = {
@@ -29,19 +46,6 @@ package_metadata = {
     'publishedDate': '9999-01-01T00:00:00Z',
     'version': '1.1',
 }
-
-
-def create_connection(database, user, password='', host='localhost', port='5432', sslmode=None):
-    """
-    Creates a connection to the database.
-
-    :returns: a database connection
-    :rtype: psycopg2.extensions.connection
-    """
-    global conn
-    if not conn or conn.closed:
-        conn = psycopg2.connect(dbname=database, user=user, password=password, host=host, port=port, sslmode=sslmode)
-    return conn
 
 
 def authenticate_gspread():
@@ -82,39 +86,43 @@ def set_spreadsheet_name(name):
 
 def list_source_ids(pattern=''):
     """
-    Returns, as a data frame, a list of source IDs matching the given pattern.
+    Returns, as a ResultSet or DataFrame, a list of source IDs matching the given pattern.
 
     :param str pattern: a substring, like "paraguay"
-    :returns: the results as a data frame
-    :rtype: pandas.DataFrame
-    """
+    :returns: the results as a pandas DataFrame or an ipython-sql `ResultSet <https://github.com/catherinedevlin/ipython-sql/blob/b24ac6e9410416eafde86ae22fd8d6f34acbe05d/src/sql/run.py#L99>`__, depending on whether ``%config SqlMagic.autopandas`` is ``True`` or ``False`` respectively. This is the same behaviour as ipython-sql's ``%sql`` magic.
+    :rtype: pandas.DataFrame or sql.run.ResultSet
+    """  # noqa: E501
     sql = """
     SELECT source_id
     FROM collection
-    WHERE source_id ILIKE %(pattern)s
+    WHERE source_id ILIKE :pattern
     GROUP BY source_id
     ORDER BY source_id
     """
 
-    return get_dataframe_from_query(sql, {'pattern': '%{}%'.format(pattern)})
+    pattern = f'%{pattern}%'
+
+    # This inspects locals to find pattern
+    return get_ipython().magic(f'sql {sql}')
 
 
 def list_collections(source_id):
     """
-    Returns, as a data frame, a list of collections with the given source ID.
+    Returns, as a ResultSet or DataFrame, a list of collections with the given source ID.
 
     :param str source_id: a source ID
-    :returns: the results as a data frame
-    :rtype: pandas.DataFrame
-    """
+    :returns: the results as a pandas DataFrame or an ipython-sql `ResultSet <https://github.com/catherinedevlin/ipython-sql/blob/b24ac6e9410416eafde86ae22fd8d6f34acbe05d/src/sql/run.py#L99>`__, depending on whether ``%config SqlMagic.autopandas`` is ``True`` or ``False`` respectively. This is the same behaviour as ipython-sql's ``%sql`` magic.
+    :rtype: pandas.DataFrame or sql.run.ResultSet
+    """  # noqa: E501
     sql = """
     SELECT *
     FROM collection
-    WHERE source_id = %(source_id)s
+    WHERE source_id = :source_id
     ORDER BY id DESC
     """
 
-    return get_dataframe_from_query(sql, {'source_id': source_id})
+    # This inspects locals to find source_id
+    return get_ipython().magic(f'sql {sql}')
 
 
 def set_search_path(schema_name):
@@ -124,70 +132,7 @@ def set_search_path(schema_name):
 
     :param str schema_name: a schema name
     """
-    with conn, conn.cursor() as cur:
-        execute_statement(cur, SQL("SET search_path = {}, public").format(Identifier(schema_name)))
-
-
-def execute_statement(cur, sql, params=None):
-    """
-    Executes a SQL statement, adding a comment with a link to the notebook for database administrators.
-
-    :param psycopg2.extensions.cursor cur: a database cursor
-    :param str sql: a SQL statement
-    :param params: the parameters to pass to the SQL statement
-    """
-    if not params:
-        params = {}
-
-    comment = '/* https://colab.research.google.com/drive/{} */'.format(_notebook_id())
-    if not isinstance(sql, str):
-        comment = SQL(comment)
-
-    try:
-        cur.execute(comment + sql, params)
-    except psycopg2.Error:
-        cur.execute('rollback')
-        raise
-
-
-def get_list_from_query(sql, params=None):
-    """
-    Executes a SQL statement and returns the results as a list of tuples.
-
-    :param str sql: a SQL statement
-    :param params: the parameters to pass to the SQL statement
-    :returns: the results as a list of tuples
-    :rtype: list
-    """
-    with conn, conn.cursor() as cur:
-        execute_statement(cur, sql, params)
-        return cur.fetchall()
-
-
-def get_dataframe_from_query(sql, params=None):
-    """
-    Executes a SQL statement and returns the results as a data frame.
-
-    :param str sql: a SQL statement
-    :param params: the parameters to pass to the SQL statement
-    :returns: the results as a data frame
-    :rtype: pandas.DataFrame
-    """
-    with conn, conn.cursor() as cur:
-        execute_statement(cur, sql, params)
-        return get_dataframe_from_cursor(cur)
-
-
-def get_dataframe_from_cursor(cur):
-    """
-    Accepts a database cursor after a SQL statement has been executed and returns the results as a data frame.
-
-    :param psycopg2.extensions.cursor cur: a database cursor
-    :returns: the results as a data frame
-    :rtype: pandas.DataFrame
-    """
-    headers = [description[0] for description in cur.description]
-    return pandas.DataFrame(cur.fetchall(), columns=headers)
+    get_ipython().magic(f'sql SET search_path = {schema_name}, public')
 
 
 def save_dataframe_to_sheet(dataframe, sheetname, prompt=True):
@@ -269,20 +214,45 @@ def download_data_as_json(data, filename):
     files.download(filename)
 
 
-def download_package_from_query(sql, params=None, package_type=None):
+def get_ipython_sql_resultset_from_query(sql):
+    """
+    Executes a SQL statement and returns a ResultSet.
+
+    Parameters are taken from the scope this function is called from (same behaviour as ipython-sql's ``%sql`` magic).
+
+    :param str sql: a SQL statement
+    :returns: the results as a `ResultSet <https://github.com/catherinedevlin/ipython-sql/blob/b24ac6e9410416eafde86ae22fd8d6f34acbe05d/src/sql/run.py#L99>`__
+    :rtype: sql.run.ResultSet
+    """  # noqa: E501
+    ipython = get_ipython()
+    autopandas = ipython.magic('config SqlMagic.autopandas')
+    # Disable autopandas, so we know that the sql magic call will always return
+    # a ResultSet (rather than a pandas DataFrame). Since the DataFrame would
+    # be created from the ResultSet, it would be less efficient.
+    if autopandas:
+        ipython.magic('config SqlMagic.autopandas=False')
+    # Use ipython.run_line_magic instead of ipython.magic here
+    # to get variables from the scope of the ipython cell,
+    # instead of the scope in this function.
+    results = ipython.run_line_magic('sql', sql)
+    if autopandas:
+        ipython.magic('config SqlMagic.autopandas=True')
+    return results
+
+
+def download_package_from_query(sql, package_type=None):
     """
     Executes a SQL statement that SELECTs only the ``data`` column of the ``data`` table, and invokes a browser
     download of the packaged data to your local computer.
 
     :param str sql: a SQL statement
-    :param params: the parameters to pass to the SQL statement
     :param str package_type: "record" or "release"
     :raises UnknownPackageTypeError: when the provided package type is unknown
     """
     if package_type not in ('record', 'release'):
         raise UnknownPackageTypeError("package_type argument must be either 'release' or 'record'")
 
-    data = [row[0] for row in get_list_from_query(sql, params)]
+    data = [row[0] for row in get_ipython_sql_resultset_from_query(sql)]
 
     if package_type == 'record':
         package = {'records': data}
@@ -310,13 +280,23 @@ def download_package_from_ocid(collection_id, ocid, package_type):
     SELECT data
     FROM data
     JOIN release ON data.id = release.data_id
-    WHERE collection_id = %(collection_id)s AND ocid = %(ocid)s
-    ORDER BY data ->> 'date' DESC
+    WHERE collection_id = :collection_id AND ocid = :ocid
+    ORDER BY data->>'date' DESC
     """
 
-    params = {'ocid': ocid, 'collection_id': collection_id}
+    ipython = get_ipython()
+    autopandas = ipython.magic('config SqlMagic.autopandas')
+    # Disable autopandas, so we know that the sql magic call will always return
+    # a ResultSet (rather than a pandas DataFrame). Since the DataFrame would
+    # be created from the ResultSet, it would be less efficient.
+    if autopandas:
+        ipython.magic('config SqlMagic.autopandas=False')
+    # This inspects locals to find ocid and collection_id
+    results = ipython.magic(f'sql {sql}')
+    if autopandas:
+        ipython.magic('config SqlMagic.autopandas=True')
 
-    data = [row[0] for row in get_list_from_query(sql, params)]
+    data = [row[0] for row in results]
 
     if package_type == 'record':
         package = {'records': [{'ocid': ocid, 'releases': data}]}